requirements.txt
flask
flask_login
flask_migrate
WTForms
flask_wtf

flask_sqlalchemy
sqlalchemy
sqlalchemy_utils
sqlalchemy_schemadisplay

email_validator
gunicorn
jinja2
flask-restx
Werkzeug
python-dotenv
Flask-Minify
Flask-Dance
blinker
pyOpenSSL

flask_admin
flask_security

stix2
pandas
numpy
neo4j
PyMySQL
mysql-connector-python
html2text
openpyxl
Flask-WTF

bleach
types-bleach

Flask-Assets
libsass
glob2


neo4j==5.14.0

python-nmap

# flask_mysqldb
# psycopg2-binary

atlassian-python-api

Flask-Mailman